done with error checking

+++ libft/ft_atoi.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aez-zaou <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: aez-zaou <aez-zaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/10 11:23:51 by aez-zaou          #+#    #+#             */
-/*   Updated: 2019/10/21 17:02:08 by aez-zaou         ###   ########.fr       */
+/*   Updated: 2021/05/21 13:19:30 by aez-zaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int		sign;
+	long	sign;
 	long	num;
 	long	max;
